move post_courses_test to testing package
